+++ changelog.md
@@ -1,4 +1,4 @@
 # 0.3.0
 
 - add `pdx.__version__`
-
+- use "FROM-first" syntax in DuckDB: https://duckdb.org/2022/11/14/announcing-duckdb-060.html